Working version of mediator
If in LED_facade and PushButton_facade the cfg struct is chagned from pointers to variables and these changes are done in uz_led_facace_init and pushbutton as well as in the mediator (assignemnt to value instead of reference) the system crashes. E.g. the expression LedRunning changes its pin number after the pushbutton are initialized