feat: Configure link-local ip address if DHCP fails
This simplifies working with the UZ board. On the host PC configure
the network interface as DHCP. If no DHCP server is available the
UZ board uses this ip and the host PC also configures automatically
an IP in this subnet range. So you can easily connect to the UZ
without changing/adding network config on host PC!